feat: Move firebaseui-web project into glitch-projects directory
This commit moves the entire `firebaseui-web` project into a new directory called `glitch-projects`. This is done to better organize the repository and group all Glitch-related projects under a single directory.

The following files and directories were moved:
- `.glitch-assets`
- `CustomElements/FireabseApp.js`
- `LICENSE`
- `README.md`
- `TODO.md`
- `app.js`
- `currentUser.html`
- `firebase-auth-config.js`
- `firebase-auth.js`
- `firebase-multi-factor-info.js`
- `firebase-options.js`
- `firebase-user-info.js`
- `firebase-user.js`
- `google.html`
- `index.html`
- `raw.html`
- `script.js`
- `style.css`

This change should not affect the functionality of the `firebaseui-web` project itself, only its location within the repository.